flask/requirements.txt
flask
gunicorn
Flask-SocketIO
eventlet
psycopg2-binary